refactor 'pytest' and 'gotest' modules into 'executors'
