+++ tests/scenarios/test_simple_scenarios.py
@@ -6,9 +6,9 @@ from pytest_mock import MockerFixture
 from step_exec_lib.types import StepType
 
 from app_test_suite.steps.base import CONTEXT_KEY_CHART_YAML, TestExecutor
-from app_test_suite.steps.pytest.pytest import PytestExecutor
+from app_test_suite.steps.executors.pytest import PytestExecutor
 from app_test_suite.steps.scenarios.simple import SmokeTestScenario, FunctionalTestScenario, TEST_APP_CATALOG_NAME
-from app_test_suite.steps.gotest.gotest import GotestExecutor
+from app_test_suite.steps.executors.gotest import GotestExecutor
 from app_test_suite.steps.scenarios.simple import SimpleTestScenario
 from tests.helpers import (
     assert_deploy_and_wait_for_app_cr,
